pyproject.toml
[project]
name = "WasteClassificationPytorch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "hydra-core>=1.3.2",
    "numpy>=2.3.1",
    "pandas>=2.3.1",
    "scikit-learn>=1.7.1",
    "tensorboard>=2.20.0",
    "torch>=2.7.1",
    "torcheval>=0.0.7",
    "torchmetrics>=1.8.0",
    "torchvision>=0.22.1",
    "tqdm>=4.67.1",
    "transformers[torch]>=4.53.3",
    "triton-windows<3.3",
]

[tool.uv]
link-mode = "copy"

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]

[dependency-groups]
dev = [
    "matplotlib>=3.10.3",
    "ruff>=0.12.5",
    "seaborn>=0.13.2",
]
